% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{is_meta}
\alias{is_meta}
\alias{as_meta}
\title{metathis Object Checks and Coercions}
\usage{
is_meta(x)

as_meta(x)
}
\arguments{
\item{x}{A list or metathis object}
}
\description{
Functions to check if an object is a \pkg{metathis} obect or to
coerce an object into a metathis object.
}
