% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlmeta-package.R
\docType{package}
\name{htmlmeta-package}
\alias{htmlmeta}
\alias{htmlmeta-package}
\title{htmlmeta: Create HTML Metadata Tags with Ease}
\description{
Creates metadata tags for use in R Markdown HTML
    documents and Shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/htmlmeta}
  \item Report bugs at \url{https://github.com/gadenbuie/htmlmeta/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (0000-0002-7111-0077)

}
\keyword{internal}
