% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_name}
\alias{meta_name}
\title{Create name/content metadata tag pairs.}
\usage{
meta_name(.meta = meta(), ...)
}
\arguments{
\item{...}{Name (argument names) and content (argument value) pairs.}
}
\description{
Creates metadata tag pairs where the arguments are the name values and their
values are content values.
}
