% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metathis-package.R
\docType{package}
\name{metathis-package}
\alias{metathis}
\alias{metathis-package}
\title{metathis: HTML Metadata Tags for R Markdown and Shiny}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Create meta tags for R Markdown HTML documents
    and Shiny apps for customized social media cards, for accessibility, and 
    quality search engine indexing. metathis currently supports HTML documents
    created with rmarkdown, shiny, xaringan, pagedown, bookdown, and 
    flexdashboard.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/metathis}
  \item Report bugs at \url{https://github.com/gadenbuie/metathis/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
