% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{include_meta}
\alias{include_meta}
\title{Include Metadata Tags in HTML Document}
\usage{
include_meta(.meta)
}
\arguments{
\item{.meta}{A \code{meta} object created by \code{\link[=meta]{meta()}} or \code{\link[=as_meta]{as_meta()}}, or returned
by a \code{meta_*()} object.}
}
\value{
An \code{\link[htmltools:htmlDependecy]{htmltools::htmlDependecy()}} containing the metadata tags to be
included in the \code{<head>} of the HTML document.
}
\description{
Use \code{include_meta()} to explicitly declare the \code{\link[=meta]{meta()}} tags as an HTML
dependency. In general, this is not required when knitting to an HTML
document. In some non-standard cases, such as with \pkg{xaringan} slides
this is necessary.
}
\seealso{
Other meta: \code{\link{meta_general}},
  \code{\link{meta_geo}}, \code{\link{meta_social}},
  \code{\link{meta_viewport}}
}
\concept{meta}
