% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metathis-package.R
\docType{package}
\name{metathis-package}
\alias{metathis}
\alias{metathis-package}
\title{metathis: Create HTML Metadata Tags with Ease}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Creates metadata tags for use in R Markdown HTML
    documents and Shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/metathis}
  \item Report bugs at \url{https://github.com/gadenbuie/metathis/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (0000-0002-7111-0077)

}
\keyword{internal}
